+++ AspNetCoreApiExample/Controllers/BlogsController.cs
@@ -11,10 +11,13 @@
 namespace Honememo.AspNetCoreApiExample.Controllers
 {
     using System.Collections.Generic;
+    using System.ComponentModel.DataAnnotations;
     using System.Linq;
     using System.Threading.Tasks;
+    using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using Honememo.AspNetCoreApiExample.Entities;
+    using Honememo.AspNetCoreApiExample.Exceptions;
     using Honememo.AspNetCoreApiExample.Repositories;
 
     /// <summary>
@@ -23,7 +26,7 @@ namespace Honememo.AspNetCoreApiExample.Controllers
     [Produces("application/json")]
     [Route("api/[controller]")]
     [ApiController]
-    public class BlogsController : ControllerBase
+    public class BlogsController : AppControllerBase
     {
         #region メンバー変数
 
@@ -72,19 +75,22 @@ namespace Honememo.AspNetCoreApiExample.Controllers
             return await this.blogRepository.FindOrFail(id);
         }
 
-        // TODO: 更新系APIには認証かける
-
         /// <summary>
         /// ブログを登録する。
         /// </summary>
-        /// <param name="blog">ブログ。</param>
+        /// <param name="body">ブログ情報。</param>
         /// <returns>登録したブログ。</returns>
         [HttpPost]
+        [Authorize]
         [ProducesResponseType(201)]
         [ProducesResponseType(400)]
-        public async Task<ActionResult<Blog>> PostBlog(Blog blog)
+        public async Task<ActionResult<Blog>> PostBlog(BlogBody body)
         {
-            await this.blogRepository.Create(blog);
+            var blog = await this.blogRepository.Create(new Blog()
+            {
+                Name = body.Name,
+                UserId = this.UserId,
+            });
             return this.CreatedAtAction(nameof(this.GetBlog), new { id = blog.Id }, blog);
         }
 
@@ -92,14 +98,21 @@ namespace Honememo.AspNetCoreApiExample.Controllers
         /// ブログを更新する。
         /// </summary>
         /// <param name="id">ブログID。</param>
-        /// <param name="blog">更新するブログ情報。</param>
+        /// <param name="body">ブログ情報。</param>
         /// <returns>処理結果。</returns>
         [HttpPut("{id}")]
+        [Authorize]
         [ProducesResponseType(204)]
         [ProducesResponseType(404)]
-        public async Task<IActionResult> PutBlog(int id, Blog blog)
+        public async Task<IActionResult> PutBlog(int id, BlogBody body)
         {
-            blog.Id = id;
+            var blog = await this.blogRepository.FindOrFail(id);
+            if (blog.UserId != this.UserId)
+            {
+                throw new ForbiddenException($"id={id} does not belong to me");
+            }
+
+            blog.Name = body.Name;
             await this.blogRepository.Update(blog);
             return this.NoContent();
         }
@@ -110,12 +123,36 @@ namespace Honememo.AspNetCoreApiExample.Controllers
         /// <param name="id">ブログID。</param>
         /// <returns>処理結果。</returns>
         [HttpDelete("{id}")]
+        [Authorize]
         [ProducesResponseType(204)]
         [ProducesResponseType(404)]
         public async Task<IActionResult> DeleteBlog(int id)
         {
+            var blog = await this.blogRepository.FindOrFail(id);
+            if (blog.UserId != this.UserId)
+            {
+                throw new ForbiddenException($"id={id} does not belong to me");
+            }
+
             await this.blogRepository.Delete(id);
             return this.NoContent();
+        }
+
+        #endregion
+
+        #region 内部クラス
+
+        /// <summary>
+        /// ブログ登録/編集のリクエストパラメータ。
+        /// </summary>
+        public class BlogBody
+        {
+            /// <summary>
+            /// ブログタイトル。
+            /// </summary>
+            [Required]
+            [MaxLength(191)]
+            public string Name { get; set; }
         }
 
         #endregion
